refactor(dot): changed the name of files and folders
