+++ src/controllers/user.controller.ts
@@ -1,7 +1,7 @@
 // Copyright 2020 Fazt Community ~ All rights reserved. MIT license.
 import { Handler } from '../types';
 import User from '../models/User';
-import { generateAndSignToken } from '../util/auth'
+import { generateAndSignToken } from '../utils/auth';
 import { ErrorHandler } from '../error';
 import { NOT_FOUND, BAD_REQUEST, UNAUTHORIZED } from 'http-status-codes';
 
@@ -31,7 +31,7 @@ export const createUser: Handler = async (req, res) => {
   await user.setPassword(password);
 
   const newUser = await user.save();
-  const token = await generateAndSignToken({ user: newUser.id })
+  const token = await generateAndSignToken({ user: newUser.id });
   console.log(newUser);
   delete newUser.password;
   return res.status(201).json({
@@ -71,7 +71,7 @@ export const signinUser: Handler = async (req, res) => {
   //compare password
   const passwordCorrect: boolean = await crendential.comparePassword(password);
   if (!passwordCorrect) {
-    throw new ErrorHandler(UNAUTHORIZED, 'Invalid Credentials')
+    throw new ErrorHandler(UNAUTHORIZED, 'Invalid Credentials');
   }
   const token = await generateAndSignToken({ user: crendential.id });
   return res.status(200).json({
