+++ src/routes/users.routes.ts
@@ -2,7 +2,7 @@
 import { Router } from 'express';
 import * as userCtrl from '../controllers/user.controller';
 import { handlerExceptionRoute } from '../error';
-import { authMiddleware } from '../middleware/auth.middleware'
+import { authMiddleware } from '../middlewares/auth.middleware';
 
 const router = Router();
 
@@ -82,7 +82,11 @@ router.get('/:id', handlerExceptionRoute(userCtrl.getUser));
  */
 router.put('/', handlerExceptionRoute(userCtrl.updateUser));
 
-router.delete('/:id', authMiddleware, handlerExceptionRoute(userCtrl.deleteUser));
+router.delete(
+  '/:id',
+  authMiddleware,
+  handlerExceptionRoute(userCtrl.deleteUser)
+);
 
 router.post('/signin', handlerExceptionRoute(userCtrl.signinUser));
 export default router;
